+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Remove unnecessary settings. Refs UIORG-150.
 * Handle "Active/Inactive" hard-code issue for the search filter. Fixes UIU-894.
 * Better search string construction won't accidentally match `*`. Fixes UIU-1086.
+* Fix sort patron blocks. Fixes UIU-868.
+* Fix Payment Status from "Charge and pay now". Fixes UIU-931.
+* Fix UX Consistency owners. Fixes UIU-1050.
 
 ## [2.23.0](https://github.com/folio-org/ui-users/tree/v2.23.0) (2019-06-12)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.22.0...v2.23.0)